add fedora and fix concurrent issue
Add fedora and refactor the code so that other distros can easily be added.

The script was using a file named cidfile to store the containment id. 2
scripts couldn't run concurrently because they wanted access to the same file.
I've instead use now a temporary file created with mktemp thus it is a different
file on every run.

I've also added an
 echo "Bye"
when exiting, just for debugging purposes